Add endpoint to accept http start requests from hm9000
  - hm9000 controller accepts both start and stop requests

[#112003111]
