+++ app/controllers/internal/dea/hm9000_controller.rb
@@ -4,7 +4,7 @@ require 'cloud_controller/internal_api'
 
 module VCAP::CloudController
   module Dea
-    class HM9000StopController < RestController::BaseController
+    class HM9000Controller < RestController::BaseController
       allow_unauthenticated_access
 
       def initialize(*)
@@ -13,6 +13,16 @@ module VCAP::CloudController
         unless auth.provided? && auth.basic? && auth.credentials == InternalApi.credentials
           raise CloudController::Errors::NotAuthenticated
         end
+      end
+
+      post '/internal/dea/hm9000/start/:app_guid', :start
+
+      def start(app_guid)
+        raise CloudController::Errors::ApiError.new_from_details('NotFound') unless App.find(guid: app_guid)
+        logger.info('dea.hm9000.start', app_guid: app_guid)
+        SubSystem.hm9000_respondent.process_hm9000_start(read_body)
+
+        [200, '{}']
       end
 
       post '/internal/dea/hm9000/stop/:app_guid', :stop
@@ -28,16 +38,16 @@ module VCAP::CloudController
       private
 
       def read_body
-        stop_message = {}
+        message = {}
         begin
           payload = body.read
-          stop_message = MultiJson.load(payload, symbolize_keys: false)
+          message = MultiJson.load(payload, symbolize_keys: false)
         rescue MultiJson::ParseError => pe
-          logger.error('dea.hm9000.stop.parse-error', payload: payload, error: pe.to_s)
+          logger.error('dea.hm9000.parse-error', payload: payload, error: pe.to_s)
           raise CloudController::Errors::ApiError.new_from_details('MessageParseError', payload)
         end
 
-        stop_message
+        message
       end
     end
   end
